// src/hub-rest-client/openapi/models/message.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Farcaster Hub REST API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import { HashScheme } from './hash-scheme';
// May contain unused imports in some cases
// @ts-ignore
import { MessageData } from './message-data';
// May contain unused imports in some cases
// @ts-ignore
import { SignatureScheme } from './signature-scheme';

/**
 * *  A Message is a delta operation on the Farcaster network. The message protobuf is an envelope  that wraps a MessageData object and contains a hash and signature which can verify its authenticity.
 * @export
 * @interface Message
 */
export interface Message {
    /**
     * 
     * @type {MessageData}
     * @memberof Message
     */
    'data': MessageData;
    /**
     * 
     * @type {string}
     * @memberof Message
     */
    'hash': string;
    /**
     * 
     * @type {HashScheme}
     * @memberof Message
     */
    'hashScheme': HashScheme;
    /**
     * 
     * @type {string}
     * @memberof Message
     */
    'signature': string;
    /**
     * 
     * @type {SignatureScheme}
     * @memberof Message
     */
    'signatureScheme': SignatureScheme;
    /**
     * 
     * @type {string}
     * @memberof Message
     */
    'signer': string;
}